feat(validation): implement comprehensive local validation pipeline to prevent CI failures
- Add LOCAL_VALIDATION_GUIDE.md with quick commands and troubleshooting
- Update README.md with local validation section and pre-commit safety guidance
- Create safe-commit.sh script for emergency bypass guidance
- Fix ESLint v9+ configuration with proper ignores for examples/** and .claude/**
- Add @ai-fastify-template/types as explicit devDependency to backend-api
- Fix module resolution errors in backend-api test imports
- Update property tests to use main package exports instead of subpath exports
- Fix Math.fround() constraints in all fast-check float generators
- Resolve security/detect-object-injection ESLint warning in property-testing-simple
- Fix floating-point precision edge cases in property tests
- Ensure all tests pass with proper error handling and boundary conditions

CRITICAL: Local validation now prevents CI failures by catching issues locally
in seconds rather than minutes in CI. Pre-commit hooks provide comprehensive
validation including security scanning, code quality, and build verification.

This completes the local/CI validation alignment objective - developers can
now confidently validate their work locally before pushing, eliminating
expensive CI failures and improving development velocity.

+++ apps/backend-api/test/utils/calculations.test.ts
@@ -124,7 +124,7 @@ describe('Property-based tests', () => {
         fc.property(
           fc.array(
             fc.record({
-              price: fc.float({ min: 0, max: 1000, noNaN: true }),
+              price: fc.float({ min: 0, max: Math.fround(1000), noNaN: true }),
               quantity: fc.integer({ min: 0, max: 100 }),
             })
           ),
@@ -141,7 +141,7 @@ describe('Property-based tests', () => {
         fc.property(
           fc.array(
             fc.record({
-              price: fc.float({ min: 0, max: 100, noNaN: true }),
+              price: fc.float({ min: 0, max: Math.fround(100), noNaN: true }),
               quantity: fc.integer({ min: 0, max: 10 }),
             }),
             { minLength: 2 }
@@ -161,7 +161,7 @@ describe('Property-based tests', () => {
         fc.property(
           fc.array(
             fc.record({
-              price: fc.float({ min: 0, max: 100, noNaN: true }),
+              price: fc.float({ min: 0, max: Math.fround(100), noNaN: true }),
               quantity: fc.integer({ min: 0, max: 10 }),
             }),
             { maxLength: 5 }
@@ -183,8 +183,8 @@ describe('Property-based tests', () => {
     it('should never return discount greater than price', () => {
       fc.assert(
         fc.property(
-          fc.float({ min: 0, max: 10000, noNaN: true }),
-          fc.float({ min: 0, max: 100, noNaN: true }),
+          fc.float({ min: 0, max: Math.fround(10000), noNaN: true }),
+          fc.float({ min: 0, max: Math.fround(100), noNaN: true }),
           (price, discountPercent) => {
             const discount = calculateDiscount(price, discountPercent);
             expect(discount).toBeLessThanOrEqual(price);
@@ -197,9 +197,9 @@ describe('Property-based tests', () => {
     it('should be monotonic (higher discount % = higher discount amount)', () => {
       fc.assert(
         fc.property(
-          fc.float({ min: 1, max: 1000, noNaN: true }),
-          fc.float({ min: 0, max: 50, noNaN: true }),
-          fc.float({ min: 0, max: 50, noNaN: true }),
+          fc.float({ min: 1, max: Math.fround(1000), noNaN: true }),
+          fc.float({ min: 0, max: Math.fround(50), noNaN: true }),
+          fc.float({ min: 0, max: Math.fround(50), noNaN: true }),
           (price, discount1, discount2) => {
             const [lower, higher] =
               discount1 <= discount2
@@ -222,11 +222,11 @@ describe('Property-based tests', () => {
         fc.property(
           fc.array(
             fc.record({
-              price: fc.float({ min: 0, max: 100, noNaN: true }),
+              price: fc.float({ min: 0, max: Math.fround(100), noNaN: true }),
               quantity: fc.integer({ min: 0, max: 10 }),
             })
           ),
-          fc.float({ min: 0, max: 1, noNaN: true }),
+          fc.float({ min: 0, max: Math.fround(1), noNaN: true }),
           (items, taxRate) => {
             const baseTotal = calculateTotal(items);
             const totalWithTax = calculateTotalWithTax(items, taxRate);
@@ -241,7 +241,7 @@ describe('Property-based tests', () => {
         fc.property(
           fc.array(
             fc.record({
-              price: fc.float({ min: 0, max: 100, noNaN: true }),
+              price: fc.float({ min: 0, max: Math.fround(100), noNaN: true }),
               quantity: fc.integer({ min: 0, max: 10 }),
             })
           ),
